US Aggression
FACT SHEET: JANUARY 2017
Human Rights First
President Trump’s Draft Executive Order on Detention and
Interrogation
On January 25, 2017 The Washington Post
publishing a draft executive order (EO) entitled
“Detention and Interrogation of Enemy
Combatants.”
1 This draft order appears to be an
attempt to pave the way to reopen CIA “black sites”
and reauthorize so-called “enhanced interrogation
techniques”2 which are illegal under both domestic
and international law.3
Among other matters, the draft EO would reinstate
Executive Order 13440 from 2007,4 which was
criticized by the military5 as well as by President
Reagan appointees General P.X. Kelley and Robert
F. Turner, for giving “the CIA carte blanche to
engage in ‘willful and outrageous acts of personal
abuse’”6
.
Secretary of Defense General James Mattis has
cited rapport-based interrogation techniques as
more effective than torture at gaining actionable
intelligence.7 Professional interrogators agree with
General Mattis that torture is ineffective,
8 as do 176
retired generals and admirals who wrote to
President Trump in January 2017, stating that
 1 https://www.washingtonpost.com/world/national-security/whitehouse-draft-order-calls-for-review-on-use-of-cia-black-sitesoverseas/2017/01/25/e4318970-e310-11e6-a547-
5fb9411d332c_story.html?utm_term=.53068b17aec9.
2 http://www.humanrightsfirst.org/sites/default/files/EnhancedInterrogation-Fact-Sheet.pdf;

http://www.humanrightsfirst.org/senate-report-cia-torture.
3 http://www.humanrightsfirst.org/sites/default/files/Issue-BriefLaws-Against-Torture.pdf.

4 https://www.gpo.gov/fdsys/pkg/FR-2007-07-24/pdf/07-3656.pdf;.
5 http://www.humanrightsfirst.org/sites/default/files/Military-citesrisk-of-abuse-by-CIA.pdf.
torture “undermines our national security.”
9 This
group of high ranking retired military leaders affirm
that torture “increases the risks to our troops,
hinders cooperation with allies, [and] alienates
populations whose support the United States needs
in the struggle against terrorism.”10
This document summarizes the provisions in the
draft EO.
þ Section 1: Revokes President Obama’s
Executive Orders 13491 on humane treatment
of detainees and 13492 on closure of the
Guantanamo Bay detention facility. Reinstates
President George W. Bush’s Executive Order
13440 “to the extent permitted by law.” Defines
torture and cruel, inhuman, and degrading
treatment in a manner that could pave the way
for their use while feigning compliance with
domestic and international legal obligations.
þ Section 2: Touts the detention facilities at
Guantanamo Bay as legal, safe, and effective;
6 http://www.washingtonpost.com/wpdyn/content/article/2007/07/25/AR2007072501881.html.

7 http://www.humanrightsfirst.org/sites/default/files/interrogatorletter-candidates.pdf.
8 https://www.theguardian.com/us-news/2017/jan/25/trumpexecutive-order-torture-black-sites-guantanamobay?CMP=share_btn_tw.
See also
http://www.humanrightsfirst.org/sites/default/files/interrogatorletter-candidates.pdf.
9 http://www.humanrightsfirst.org/sites/default/files/torturegenerals-admirals-letter-to-trump.pdf.
10 Id.
Fact Sheet: 2017
Human Rights First
recites misleading statistics on rates of released
detainees engaging in future terrorist acts;11 and
finds that it is in U.S. interests to continue
detention operations at the base.
þ Section 3: Finds that the United States remains
engaged in broadly defined “global armed
conflict” with not only members al Qaeda, the
Taliban, and associated forces—including
members of the Islamic State in Iraq and Syria—
but also those accused of fighting on behalf of
these groups or providing “substantial support”
to them in furtherance of hostilities against the
United States or its partners.
þ Section 4: Subject to “further direction from the
President,” orders the secretary of defense to
continue detention operations and military
commission trials of alien enemy combatants
captured in the global conflict as that conflict is
defined in Section 3, including newly captured
individuals “as appropriate.” Orders such
operations to be consistent with the law.
Suspends any detainee transfer efforts that are
underway pending further review. Reserves the
authority to hold detainees in “other facilities
available to the United States[.]”
þ Section 5: Orders the secretary of defense to
review the interrogation policies in the Army
Field Manual and make changes and additions,
which must be consistent with the law, to ensure
“the safe, lawful, and effective interrogation of
enemy combatants” captured in what it calls the
“fight against radical Islamism.”
12
þ Section 6: Orders the attorney general to review
Presidential Policy Directive 14 (PPD 14),
13
which governs whether newly captured
individuals will be held in law enforcement or
military custody. Asks the attorney general to
 11 http://www.humanrightsfirst.org/sites/default/files/GuantanamoDetainee-Transfers-Fact-Sheet.pdf.
12 As in other sections, the phrase “global war on terrorism” is
shown as crossed out and replaced with this term.
recommend any changes to PPD 14 that are
“necessary, lawful, and appropriate” within 120
days of the date the EO goes into effect.
þ Section 7: Orders the director of national
intelligence to review current intelligence needs
of the United States “in the fight against radical
Islamism” and to recommend: a) whether to
reinstate CIA-run detention facilities outside of
the United States; b) whether to modify EO
13440 that was reinstated in Section 1; and c)
whether any legislative changes are necessary
to “permit the resumption of an effective and
lawful interrogation program.” Notably, the
explanatory statement of the EO says that
Congress has enacted restrictions that provide
“a significant statutory barrier” to the resumption
of the CIA interrogation program. It also laments
the statutory requirement that any changes to
the list of approved interrogation techniques
must be made public, a requirement that was
designed to prevent future administrations from
secretly authorizing techniques that constitute
torture or other cruel treatment.
þ Section 8: Orders the secretary of defense to
review the military commission system in use at
Guantanamo Bay and recommend how to
employ the system going forward to prosecute
and punish “unlawful enemy combatants
detained in the armed conflict with violent
Islamic extremists.”14
þ Section 9: Orders that no person in the custody
of the United States may be subjected to torture
or cruel, inhuman, or degrading treatment or
punishment but only “as proscribed by U.S. law.”
Further orders that the detention, treatment,
interrogation, and transfer of enemy combatants
comply with “all the laws of the United States.”
13 https://obamawhitehouse.archives.gov/the-pressoffice/2012/02/28/presidential-policy-directive-requirementsnational-defense-authorizatio.
14 This is the lone use of the term “Islamic extremists” in the entire
EO.
Fact Sheet: 2017
Human Rights First
This section purports to require compliance with
the law, including the prohibition on torture.
However, other provisions in the EO and its
accompanying explanatory statement suggest
the new administration may attempt to interpret
or revise existing law to justify a return to
unlawful and abusive practices. Such “loophole
lawyering” was used during the George W. Bush
Administration to secretly permit unlawful
treatment of detainees while publicly claiming
that it was not engaged in torture.
þ Section 10: Contains generic legal disclaimers,
including, notably, that the order must be
implemented consistent with applicable law and
subject to available appropriations.